refactor: move od validation types from jobDetails.ts to job.validation.ts
